FIX: support for multi certificates
We introduced that option recently but I don't think we did it right.

The current setup is assuming that `:idp_cert` is our primary certificate and `:idp_cert_multi` is secondary.

We are using [omniauth-saml](https://github.com/omniauth/omniauth-saml) gem in background and this get is using [ruby-saml](https://github.com/onelogin/ruby-saml)

According to `ruby-saml` documentation:

```
In order to handle that the toolkit offers the 'idp_cert_multi' parameter. When used, 'idp_cert' and 'idp_cert_fingerprint' values are ignored.

That 'idp_cert_multi' must be a Hash as follows: { :signing => [], :encryption => [] }
```

It means that we need an ability to define 2 certificates as `multi` and merge it to config only if the value is provided.